Revert "Allow wildcard patterns in matches"
This reverts commit aba72d340d5fd729790c3ec3aa01cd5a4d8221e7.

Turns out that adding in `wildcard_pat` wasn't necessary, since "_" is
accepted as a valid ident_pat already, which is consistent with Rust's
definition here: https://doc.rust-lang.org/reference/identifiers.html

Thanks @parno for noticing this.

The original reason for adding this commit was as a fix for an issue I
was seeing, but that seems to have been fixed up by a later commit, so
this is no longer necessary.