Separate out quantifier expressions from closures
This is primarily so that we can have the `expr_no_struct` be there for
a `forall` or such, but use `expr` for "normal" closures.

This split is necessary because otherwise, this gets parsed incorrectly:

  proof fn foo()
    ensures forall|x:int| x == x
  {
  }

Notice the lack of `,` there; that causes the parser to believe that the
`x` on the rhs of the `==` to be considered as a struct, which is
incorrect.