Allow code that just has a `verus!{`
Notice that there is a lack of space between `verus!` and `{`. This was
not an issue before, when we ran rustfmt before the main verusfmt
formatter. However, with #23, we changed the ordering to help with
prettier error reports. Nonetheless, without rustfmt jumping in and
adding in that space, the parser would not dive into anything that was a
`verus!{...`, and thus would not format anything.

Turns out, due to this exact thing, some tests in our test suite were
not even running; thus they started to fail once the issue was fixed.
I've fixed up these tests now (in the same commit)