A new approach to handling non-inline single-line comments (#28)
During formatting, rather than trying to dynamically decide when a single-line comment should be placed on its own line based on its context, when we encounter a single-line comment that was not originally inline, we mark it as such, and we also add a destination marker on the next line. After everything is formatted, we can scan for the non-inline markers. If an originally non-inline comment has been inlined, we move it to its corresponding destination marker. We need the destination marker to tells us how much the comment should be indented when moved to the next line.