Fix incorrect parsing of string contains "verus!{".
non_verus is defined to accept ANY which might be of low precedency.
When a string literal in non_verus contains "verus!{", VerusParser fails to parse the string and throws an error indicating the string is parsed into verus_macro_use.

This patch modfies non_verus so it accepts string, raw_string, etc., explicitly so string literals with "verus!{" can be correctly parsed.

verus-minimal is changed as well otherwise rustfmt shipped in this crate cannot format the non-verus code correctly.

A companion unit test which should fail without this patch is included.