Support older powershell
Updates pre-built binary installation instructions to support both PowerShell 5.x (default on Windows 11), as well as latest (7.x). Prior instructions were tested on 7.x. As far as I can tell, 6.x onwards is when things changed to the new behavior that doesn't need to set the ExecutionPolicy. Nonetheless, the updated instructions have now been tested to work on both 5.1.22621.3880 and 7.4.3.