Support space between `verus !` for macro
While _technically_ this is not necessary, it does make it _much_ easier
to deal with `creduce`d files, which otherwise easily go and hit this
particular low-hanging issue, rather than actually interesting issues.

Fwiw, `rustfmt` automatically collapses any `macro !` to `macro!`, so we
would not be wrong in handling this specific extra amount of whitespace
during parsing.